* add namespaceAppend
