+++ src/components/EntitlementsAgreementsList/index.js
@@ -1,2 +1,1 @@
 export { default } from './EntitlementAgreementsList';
-export { default as EntitlementAgreementsBaseList } from './EntitlementAgreementsBaseList';
